pallet-revive: pluggable JIT execution backend (+ node, runtime, CI wiring)
Add a JIT execution path alongside the interpreter, sharing the same
PolkaVM module surface and host-function dispatch but selectable per
contract at execution time. Includes the surrounding node/runtime/CI
infrastructure needed for pallet-revive to consume the JIT backend.

Pallet-revive:
- Add `JitInstance` as a second `pvm::Backend` consumer of
  sp_virtualization; backend selection happens at `execute()` via
  `pvm::is_jit()` (with a bench-only `with_jit_override` for forcing
  either side).
- Defer Module compilation to `PreparedCall::new_jit` so
  `ContractBlob::code` carries raw bytes for both backends and the
  cold/warm JIT cache lookup happens once per call.
- Store `PristineCode` raw (drop the `Compact<u32>(len)++bytes`
  StorageMap wrapper) and expose it via a `pristine_code` runtime API;
  thread `Option<Vec<u8>>` through `Executable::code` so the JIT
  load-from-storage-key shortcut can skip the bytes copy.
- Token-only metering: split `WeightBackend` into a
  `PolkaVmWeightBackend` plus `BackendCosts<B>` Token wrappers, with
  Interpreter/Jit/Evm markers selecting per-instruction and per-call
  weights; set the JIT per-fuel weight to 1 cycle @ 2.6 GHz
  (hardcoded, not a placeholder).
- Gate JIT at compile time with `cfg!(revive_jit)` inlined at the
  single `from_storage` call site so host-function call sites are
  structurally dead on no-cfg builds and stripped from WASM.
- Syscall dispatch: zero-copy `resolve_import` returning `&[u8]`,
  per-syscall sentinel `u32` ids, integer-match dispatch instead of
  per-Ecalli symbol copy.
- Add benches: `seal_call_jit`, `noop_host_fn_jit`,
  `call_with_pvm_jit_{cold,warm}_cache_per_byte`; tighten JIT load
  weight via lookup-first warm-up tests; gate `pristine_code::exists`
  to bench/test cfgs.
- Structural cleanup: gather JIT-only code into `vm::pvm::jit`, drop
  pub fields/constructors, demote cfg-gated rustdoc links, narrow
  by-value passes to by-reference.

sc-virtualization:
- Expose `ExtensionsFactory` and a `default_extension` helper so node
  binaries can register the extension without reaching into manager
  internals.

Nodes / runtimes / omni-bencher:
- Declare `revive_jit` in the workspace lint config.
- Forward `--cfg revive_jit` into runtime WASM builds from each
  consuming runtime's build.rs (asset-hub-westend, kitchensink,
  revive-dev), since wasm-builder strips outer RUSTFLAGS.
- Register `sp_virtualization::HostFunctions` and
  `sc_virtualization::ExtensionsFactory` on the substrate node CLI and
  polkadot-omni-node under cfg(revive_jit); omni-bencher and the
  benchmarking-cli pull in the host functions unconditionally because
  the pristine-code runtime API uses them.
- Refresh asset-hub-westend's local pallet_revive WeightInfo override
  with stubs for new bench methods.

CI:
- Run pallet-revive benchmark and revive-touching jobs in privileged
  Docker containers so the JIT sandbox can create user namespaces.
- Add a CI workflow that runs pallet-revive tests with `--cfg
  revive_jit`, retries disabled, no fail-fast.

+++ substrate/frame/revive/src/evm/api/debug_rpc_types.rs
@@ -585,7 +585,7 @@ pub enum ExecutionStepKind {
 			serialize_with = "serialize_syscall_op",
 			deserialize_with = "deserialize_syscall_op"
 		)]
-		op: u8,
+		op: u32,
 		/// The syscall arguments (register values a0-a5).
 		/// Omitted when `disable_syscall_details` is true in ExecutionTracerConfig.
 		#[serde(default, skip_serializing_if = "Vec::is_empty", with = "super::hex_serde::vec")]
@@ -795,7 +795,7 @@ where
 }
 
 /// Serialize a syscall index to its name
-fn serialize_syscall_op<S>(idx: &u8, serializer: S) -> Result<S::Ok, S::Error>
+fn serialize_syscall_op<S>(idx: &u32, serializer: S) -> Result<S::Ok, S::Error>
 where
 	S: serde::Serializer,
 {
@@ -818,7 +818,7 @@ where
 }
 
 /// Deserialize syscall from string name to index
-fn deserialize_syscall_op<'de, D>(deserializer: D) -> Result<u8, D::Error>
+fn deserialize_syscall_op<'de, D>(deserializer: D) -> Result<u32, D::Error>
 where
 	D: serde::Deserializer<'de>,
 {
@@ -827,7 +827,7 @@ where
 	let ops = list_trace_ops();
 	ops.iter()
 		.position(|name| core::str::from_utf8(name).unwrap_or_default() == s)
-		.map(|i| i as u8)
+		.map(|i| i as u32)
 		.ok_or_else(|| serde::de::Error::custom(alloc::format!("Unknown trace op: {}", s)))
 }
